refactor(state): selectie en klembord naar een eigen slice
K-item 35 uit docs/onderhoudbaarheid/README.md §5, tweede helft. `taskSlice.ts`
gaat van 1104 naar 909 regels; de nieuwe `selectionSlice.ts` is er 227.

Selectie en klembord horen bij elkaar en niet bij de taakmutaties: het klembord
werkt op de selectie, en beide zijn iets anders dan de taken zelf. Wat ze deelt
is de UNDO-regel -- selectie is GEEN documentdata en pusht dus nooit een
snapshot, terwijl plakken dat juist wél doet omdat er taken ontstaan. Die regel
stond eerder als losse commentaarregels verspreid tussen 1100 regels
taakmutaties; nu is hij de reden dat het bestand bestaat en staat hij in de kop.

Geen gedragswijziging: de acties zijn ongewijzigd verplaatst. De ENIGE
inhoudelijke wijziging is dat `TaskClipboard` niet langer twee keer gedeclareerd
staat.

Het documentcontract verandert bewust NIET mee. `selectedTaskIds` en
`taskClipboard` blijven op het top-level van `AppState` staan -- alle slices
spreiden in één plat state-object -- dus `documentContract.ts` blijft
`s.selectedTaskIds` lezen. Dat is de reden dat deze verplaatsing goedkoop is:
had de state zich mee verplaatst naar een genest object, dan hadden de
descriptors, de undo-snapshot en de documentwissel alle drie mee gemoeten.

Geverifieerd: `npm run verify` groen (exit 0, acht stappen, verify:cycles over
383 modules). `check-document-contract.ts` blijft op 222 checks -- die poort
bewaakt precies dat selectie nog steeds per document reist en niet in een
undo-stap belandt. Met een negatieve controle bevestigd dat die poort rood wordt
zodra `selectedTaskIds` niet meer meereist.
